moved perf_metrics collector under ebpf/telemetry package
