+++ pkg/network/usm/ebpf_main.go
@@ -250,7 +250,7 @@ func (e *ebpfProgram) Close() error {
 		return nil
 	}
 	e.executePerProtocol(e.enabledProtocols, "stop", stopProtocolWrapper, nil)
-	ddebpf.UnregisterTelemetry(e.Manager.Manager)
+	ebpftelemetry.UnregisterTelemetry(e.Manager.Manager)
 	return e.Stop(manager.CleanAll)
 }
 
@@ -399,7 +399,7 @@ func (e *ebpfProgram) init(buf bytecode.AssetReader, options manager.Options) er
 	if e.cfg.InternalTelemetryEnabled {
 		for _, pm := range e.PerfMaps {
 			pm.TelemetryEnabled = true
-			ddebpf.ReportPerfMapTelemetry(pm)
+			ebpftelemetry.ReportPerfMapTelemetry(pm)
 		}
 	}
