+++ pkg/security/probe/eventstream/reorderer/perfmap.go
@@ -19,7 +19,7 @@ import (
 	manager "github.com/DataDog/ebpf-manager"
 	"github.com/cilium/ebpf/perf"
 
-	ddebpf "github.com/DataDog/datadog-agent/pkg/ebpf"
+	ebpfTelemetry "github.com/DataDog/datadog-agent/pkg/ebpf/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/security/probe/config"
 	"github.com/DataDog/datadog-agent/pkg/security/probe/eventstream"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/model"
@@ -54,7 +54,7 @@ func (m *OrderedPerfMap) Init(mgr *manager.Manager, config *config.Config) error
 		m.perfMap.PerfMapOptions.PerfRingBufferSize = config.EventStreamBufferSize
 	}
 
-	ddebpf.ReportPerfMapTelemetry(m.perfMap)
+	ebpfTelemetry.ReportPerfMapTelemetry(m.perfMap)
 	return nil
 }
